feat: move test for exact components folder + add utility components
